Add font awesome & reset functionality
