+++ src/index.html
@@ -66,11 +66,11 @@
       
       <label>priority
         
-        <input type="radio" id="high" name="priority" value="high">
+        <input type="radio" id="high" name="priority"  class="radio-button" value="high">
         <label for="high">High</label>
-        <input type="radio" id="medium" name="priority" value="medium">
+        <input type="radio" id="medium" name="priority" class="radio-button"  value="medium">
         <label for="medium">Medium</label>
-        <input type="radio" id="low" name="priority" value="low">
+        <input type="radio" id="low" name="priority" class="radio-button"  value="low">
         <label for="low">Low</label>
       </label>
       
@@ -86,6 +86,6 @@
   <div class="task-created-alert">
   </div>
   
-  <script src="index.js"></script>
+  <script src="../dist/bundle.js"></script>
 </body>
 </html>
